requirements.txt
asgiref==3.11.1
crispy-bootstrap5==2026.3
Django==6.0.3
django-crispy-forms==2.6
mysqlclient==2.2.8
sqlparse==0.5.5
tzdata==2026.1
